Refactor sidechain top pool executor (#517)
* replace previous_state_hash with more appropriate initial_state_hash

* comments only

* fix compile error

* some prepration renamings

* add get_fn to top pool operators

* clean up comments

* add remove_call & getters from pool fn
